requirements.txt
langchain
langchain-community
langchain-google-genai
google-generativeai
PyPDF2
flask
python-dotenv
faiss-cpu